+++ interfaces/yearn/DelegatedVault.sol
@@ -2,10 +2,16 @@ pragma solidity ^0.5.16;
 
 interface DelegatedVault {
     function token() external view returns (address);
-    function deposit(uint) external;
+
+    function deposit(uint256) external;
+
     function depositAll() external;
-    function withdraw(uint) external;
+
+    function withdraw(uint256) external;
+
     function withdrawAll() external;
-    function getPricePerFullShare() external view returns (uint);
+
+    function getPricePerFullShare() external view returns (uint256);
+
     function claimInsurance() external;
 }
